add risc0_verifier_router to test-utils
