+++ Cargo.toml
@@ -53,6 +53,7 @@ risc0-aggregation = { version = "0.3.0" }
 risc0-binfmt = "~1.2"
 risc0-build = { version = "~1.2", features = ["docker"] }
 risc0-build-ethereum = { version = "~1.4" }
+risc0-circuit-recursion = { version = "~1.2" }
 risc0-ethereum-contracts = { version = "~1.4" }
 risc0-zkvm = { version = "~1.2", default-features = false }
 risc0-zkp = { version = "~1.2" }
