+++ crates/boundless-cli/src/bin/boundless-cli.rs
@@ -829,7 +829,10 @@ mod tests {
     use super::*;
 
     use alloy::node_bindings::Anvil;
-    use boundless_market::contracts::{hit_points::default_allowance, test_utils::create_test_ctx};
+    use boundless_market::contracts::{
+        hit_points::default_allowance,
+        test_utils::{create_test_ctx, DEV_MODE_TRUE},
+    };
     use guest_assessor::ASSESSOR_GUEST_ID;
     use guest_set_builder::SET_BUILDER_ID;
     use tokio::time::timeout;
@@ -840,8 +843,9 @@ mod tests {
     async fn test_deposit_withdraw() {
         // Setup anvil
         let anvil = Anvil::new().spawn();
-
-        let ctx = create_test_ctx(&anvil, SET_BUILDER_ID, ASSESSOR_GUEST_ID).await.unwrap();
+        let ctx = create_test_ctx(&anvil, SET_BUILDER_ID, ASSESSOR_GUEST_ID, DEV_MODE_TRUE)
+            .await
+            .unwrap();
 
         let mut args = MainArgs {
             rpc_url: anvil.endpoint_url(),
@@ -869,8 +873,9 @@ mod tests {
     async fn test_submit_request() {
         // Setup anvil
         let anvil = Anvil::new().spawn();
-
-        let ctx = create_test_ctx(&anvil, SET_BUILDER_ID, ASSESSOR_GUEST_ID).await.unwrap();
+        let ctx = create_test_ctx(&anvil, SET_BUILDER_ID, ASSESSOR_GUEST_ID, DEV_MODE_TRUE)
+            .await
+            .unwrap();
         ctx.prover_market
             .deposit_stake_with_permit(default_allowance(), &ctx.prover_signer)
             .await
