+++ crates/boundless-market/Cargo.toml
@@ -35,6 +35,7 @@ hex = { workspace = true }
 rand = { workspace = true }
 reqwest = { workspace = true, features = ["json", "multipart"] }
 risc0-aggregation = { workspace = true }
+risc0-circuit-recursion = { workspace = true, optional = true }
 risc0-ethereum-contracts = { workspace = true, features = ["unstable"] }
 risc0-zkvm = { workspace = true, features = ["std", "client"] }
 rmp-serde = { workspace = true }
@@ -61,4 +62,4 @@ serde_json = { workspace = true }
 
 [features]
 default = []
-test-utils = []
+test-utils = ["dep:risc0-circuit-recursion"]
