+++ crates/boundless-market/src/contracts/boundless_market.rs
@@ -1349,7 +1349,7 @@ mod tests {
     use crate::{
         contracts::{
             hit_points::default_allowance,
-            test_utils::{create_test_ctx, TestCtx},
+            test_utils::{create_test_ctx, TestCtx, DEV_MODE_TRUE},
             AssessorJournal, AssessorReceipt, Fulfillment, IBoundlessMarket, Input, InputType,
             Offer, Predicate, PredicateType, ProofRequest, ProofStatus, Requirements,
         },
@@ -1527,7 +1527,9 @@ mod tests {
         // Setup anvil
         let anvil = Anvil::new().spawn();
 
-        let ctx = create_test_ctx(&anvil, SET_BUILDER_ID, ASSESSOR_GUEST_ID).await.unwrap();
+        let ctx = create_test_ctx(&anvil, SET_BUILDER_ID, ASSESSOR_GUEST_ID, DEV_MODE_TRUE)
+            .await
+            .unwrap();
 
         // Deposit prover balances
         ctx.prover_market.deposit(parse_ether("2").unwrap()).await.unwrap();
@@ -1553,7 +1555,9 @@ mod tests {
         // Setup anvil
         let anvil = Anvil::new().spawn();
 
-        let mut ctx = create_test_ctx(&anvil, SET_BUILDER_ID, ASSESSOR_GUEST_ID).await.unwrap();
+        let mut ctx = create_test_ctx(&anvil, SET_BUILDER_ID, ASSESSOR_GUEST_ID, DEV_MODE_TRUE)
+            .await
+            .unwrap();
 
         let deposit = U256::from(10);
 
@@ -1604,7 +1608,9 @@ mod tests {
         // Setup anvil
         let anvil = Anvil::new().spawn();
 
-        let ctx = create_test_ctx(&anvil, SET_BUILDER_ID, ASSESSOR_GUEST_ID).await.unwrap();
+        let ctx = create_test_ctx(&anvil, SET_BUILDER_ID, ASSESSOR_GUEST_ID, DEV_MODE_TRUE)
+            .await
+            .unwrap();
 
         let request = new_request(1, &ctx).await;
 
@@ -1624,7 +1630,9 @@ mod tests {
         // Setup anvil
         let anvil = Anvil::new().spawn();
 
-        let ctx = create_test_ctx(&anvil, SET_BUILDER_ID, ASSESSOR_GUEST_ID).await.unwrap();
+        let ctx = create_test_ctx(&anvil, SET_BUILDER_ID, ASSESSOR_GUEST_ID, DEV_MODE_TRUE)
+            .await
+            .unwrap();
 
         let eip712_domain = eip712_domain! {
             name: "IBoundlessMarket",
@@ -1700,7 +1708,9 @@ mod tests {
         // Setup anvil
         let anvil = Anvil::new().spawn();
 
-        let ctx = create_test_ctx(&anvil, SET_BUILDER_ID, ASSESSOR_GUEST_ID).await.unwrap();
+        let ctx = create_test_ctx(&anvil, SET_BUILDER_ID, ASSESSOR_GUEST_ID, DEV_MODE_TRUE)
+            .await
+            .unwrap();
 
         let eip712_domain = eip712_domain! {
             name: "IBoundlessMarket",
@@ -1782,7 +1792,9 @@ mod tests {
         // Setup anvil
         let anvil = Anvil::new().spawn();
 
-        let ctx = create_test_ctx(&anvil, SET_BUILDER_ID, ASSESSOR_GUEST_ID).await.unwrap();
+        let ctx = create_test_ctx(&anvil, SET_BUILDER_ID, ASSESSOR_GUEST_ID, DEV_MODE_TRUE)
+            .await
+            .unwrap();
 
         let eip712_domain = eip712_domain! {
             name: "IBoundlessMarket",
@@ -1853,7 +1865,9 @@ mod tests {
         // Setup anvil
         let anvil = Anvil::new().spawn();
 
-        let ctx = create_test_ctx(&anvil, SET_BUILDER_ID, ASSESSOR_GUEST_ID).await.unwrap();
+        let ctx = create_test_ctx(&anvil, SET_BUILDER_ID, ASSESSOR_GUEST_ID, DEV_MODE_TRUE)
+            .await
+            .unwrap();
 
         let eip712_domain = eip712_domain! {
             name: "IBoundlessMarket",
