+++ crates/boundless-market/src/contracts/bytecode.rs
@@ -35,3 +35,18 @@ alloy::sol! {
         constructor(address implementation, bytes memory data) payable {}
     }
 }
+
+alloy::sol! {
+    #[sol(rpc, bytecode = "60803460c157601f61082638819003918201601f19168301916001600160401b0383118484101760c55780849260209460405283398101031260c157516001600160a01b0381169081900360c157801560ae57600180546001600160a01b03199081169091555f80549182168317815560405192916001600160a01b0316907f8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e09080a361074c90816100da8239f35b631e4fbdf760e01b5f525f60045260245ffd5b5f80fd5b634e487b7160e01b5f52604160045260245ffdfe60806040526004361015610011575f80fd5b5f803560e01c80631599ead5146104fa5780633cadf449146104d6578063715018a61461047157806379ba5097146103ec5780638da5cb5b146103c557806393d237f61461034d578063ab750e7514610290578063c17e7ba814610242578063d0a6af3014610182578063e30c397814610159578063f2fde38b146100e25763fa9381d31461009e575f80fd5b346100df5760203660031901126100df576020906001600160e01b03196100c3610609565b16815260028252604060018060a01b0391205416604051908152f35b80fd5b50346100df5760203660031901126100df576004356001600160a01b0381169081900361015557610111610719565b600180546001600160a01b0319168217905581546001600160a01b03167f38d16b8cac22d99fc7c124b9cd0de2d3fa1faef420bfe791d8c362d765e227008380a380f35b5080fd5b50346100df57806003193601126100df576001546040516001600160a01b039091168152602090f35b50346100df5760403660031901126100df5761019c610609565b6024356001600160a01b038116919082900361023e576101ba610719565b6001600160e01b031916808352600260205260408320546001600160a01b031660011461022c57808352600260205260408320546001600160a01b031661021a57825260026020526040822080546001600160a01b031916909117905580f35b6350a41efb60e01b8352600452602482fd5b633780eb1360e21b8352600452602482fd5b8280fd5b50346100df5760203660031901126100df57600435906001600160401b0382116100df57602061027e6102783660048601610620565b906106fa565b6040516001600160a01b039091168152f35b50346100df5760603660031901126100df57806004356001600160401b03811161034a576102c2903690600401610620565b6001600160a01b036102d482846106fa565b1691823b156103455760405163ab750e7560e01b815260606004820152928492849283918291610308916064840191610682565b6024356024830152604435604483015203915afa801561033a576103295750f35b816103339161064d565b6100df5780f35b6040513d84823e3d90fd5b505050fd5b50fd5b50346100df5760203660031901126100df57610367610609565b61036f610719565b6001600160e01b031916808252600260205260408220546001600160a01b0316156103b357815260026020526040812080546001600160a01b031916600117905580f35b63727532a160e11b8252600452602490fd5b50346100df57806003193601126100df57546040516001600160a01b039091168152602090f35b50346100df57806003193601126100df57600154336001600160a01b039091160361045e57600180546001600160a01b0319908116909155815433918116821783556001600160a01b03167f8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e08380a380f35b63118cdaa760e01b815233600452602490fd5b50346100df57806003193601126100df5761048a610719565b600180546001600160a01b03199081169091558154908116825581906001600160a01b03167f8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e08280a380f35b50346100df5760203660031901126100df57602061027e6104f5610609565b6106a2565b5034610605576020366003190112610605576004356001600160401b038111610605578036039060406003198301126106055760048101359160221901821215610605578181016004810135906001600160401b03821161060557602401908036038213610605576001600160a01b039161057591906106fa565b1690813b15610605578060405193631599ead560e01b85526020600486015201916024600484013593016001600160401b0384116106055783360381136106055784809360246105d283945f986040848701526064860191610682565b910135604483015203915afa80156105fa576105ec575080f35b6105f891505f9061064d565b005b6040513d5f823e3d90fd5b5f80fd5b600435906001600160e01b03198216820361060557565b9181601f84011215610605578235916001600160401b038311610605576020838186019501011161060557565b90601f801991011681019081106001600160401b0382111761066e57604052565b634e487b7160e01b5f52604160045260245ffd5b908060209392818452848401375f828201840152601f01601f1916010190565b6001600160e01b0319165f818152600260205260409020546001600160a01b03169081156106e857600182146106d6575090565b633780eb1360e21b5f5260045260245ffd5b63727532a160e11b5f5260045260245ffd5b906004116106055761071690356001600160e01b0319166106a2565b90565b5f546001600160a01b0316330361072c57565b63118cdaa760e01b5f523360045260245ffdfea164736f6c634300081a000a")]
+    contract RiscZeroVerifierRouter {
+        constructor(address owner) {}
+            function addVerifier(bytes4 selector, address verifier) {}
+    }
+}
+
+alloy::sol! {
+    #[sol(rpc, bytecode = "6101808060405234610c9257604081611efd80380380916100208285610c96565b833981010312610c925780516020918201519091600883811c7eff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff169084901b7fff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff001617601081811c7dffff0000ffff0000ffff0000ffff0000ffff0000ffff0000ffff0000ffff1691901b7fffff0000ffff0000ffff0000ffff0000ffff0000ffff0000ffff0000ffff0000161780821c7bffffffff00000000ffffffff00000000ffffffff00000000ffffffff16911b7fffffffff00000000ffffffff00000000ffffffff00000000ffffffff000000001617604081811c77ffffffffffffffff0000000000000000ffffffffffffffff1691901b7fffffffffffffffff0000000000000000ffffffffffffffff00000000000000001617608081811c91901b176001600160801b031981811660a052608091821b16905260c08190526040517f72697363302e47726f74683136526563656970745665726966696572506172618152656d657465727360d01b602082810191909152905f9060269060025afa15610b11575f5190600881811c7eff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff1691901b7fff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff00ff001617601081811c7dffff0000ffff0000ffff0000ffff0000ffff0000ffff0000ffff0000ffff1691901b7fffff0000ffff0000ffff0000ffff0000ffff0000ffff0000ffff0000ffff00001617602081811c7bffffffff00000000ffffffff00000000ffffffff00000000ffffffff1691901b7fffffffff00000000ffffffff00000000ffffffff00000000ffffffff000000001617604081811c77ffffffffffffffff0000000000000000ffffffffffffffff1691901b7fffffffffffffffff0000000000000000ffffffffffffffff00000000000000001617608081811c91901b179160e0604051916103068284610c96565b60068352601f19820136602085013760205f604051828101907f12ac9a25dcd5e1a832a9061a082c15dd1d61aa9c4d553505739d0f5d65dc3be482527f025aa744581ebe7ad91731911c898569106ff5a2d30f3eee2b23c60ee980acd4604082015260408152610377606082610c96565b604051918291518091835e8101838152039060025afa15610b11575f5161039d84610ccd565b5260205f604051828101907f0707b920bc978c02f292fae2036e057be54294114ccc3c8769d883f688a1423f82527f2e32a094b7589554f7bc357bf63481acd2d55555c203383782a4650787ff6642604082015260408152610400606082610c96565b604051918291518091835e8101838152039060025afa15610b11575f5161042684610cda565b5260205f604051828101907f0bca36e2cbe6394b3e249751853f961511011c7148e336f4fd974644850fc34782527f2ede7c9acf48cf3a3729fa3d68714e2a8435d4fa6db8f7f409c153b1fcdf9b8b604082015260408152610489606082610c96565b604051918291518091835e8101838152039060025afa15610b11575f51835160021015610b5257606084015260205f604051828101907f1b8af999dbfbb3927c091cc2aaf201e488cbacc3e2c6b6fb5a25f9112e04f2a782527f2b91a26aa92e1b6f5722949f192a81c850d586d81a60157f3e9cf04f679cccd6604082015260408152610517606082610c96565b604051918291518091835e8101838152039060025afa15610b11575f51835160031015610b5257608084015260205f604051828101907f2b5f494ed674235b8ac1750bdfd5a7615f002d4a1dcefeddd06eda5a076ccd0d82527f2fe520ad2020aab9cbba817fcbb9a863b8a76ff88f14f912c5e71665b2ad5e826040820152604081526105a5606082610c96565b604051918291518091835e8101838152039060025afa15610b11575f51835160041015610b525760a084015260205f604051828101907f0f1c3c0d5d9da0fa03666843cde4e82e869ba5252fce3c25d5940320b1c4d49382527f214bfcff74f425f6fe8c0d07b307482d8bc8bb2f3608f68287aa01bd0b69e809604082015260408152610633606082610c96565b604051918291518091835e8101838152039060025afa15610b11575f51835160051015610b525760c084015260205f601a6040517f72697363305f67726f746831362e566572696679696e674b6579000000000000815260025afa15610b11575f519460205f604051828101907f2d4d9aa7e302d9df41749d5507949d05dbea33fbb16c643b22f599a2be6df2e282527f14bedd503c37ceb061d8ec60209fe345ce89830a19230301f076caff004d19266040820152604081526106f8606082610c96565b604051918291518091835e8101838152039060025afa15610b11575f519460205f604051828101907f0967032fcbf776d1afc985f88877f182d38480a653f2decaa9794cbc3bf3060c82527f0e187847ad4c798374d0d6732bf501847dd68bc0e071241e0213bc7fc13db7ab60408201527f304cfbd1e08a704a99f5e847d93f8c3caafddec46b7a0d379da69a4d112346a760608201527f1739c1b1a457a8c7313123d24d2f9192f896b7c63eea05a9d57f06547ad0cec86080820152608081526107c460a082610c96565b604051918291518091835e8101838152039060025afa15610b11575f519560205f604051828101907f198e9393920d483a7260bfb731fb5d25f1aa493335a9e71297e485b7aef312c282527f1800deef121f1e76426a00665e5c4479674322d4f75edadd46debd5cd992f6ed60408201527f090689d0585ff075ec9e99ad690c3395bc4b313370b38ef355acdadcd122975b60608201527f12c85ea5db8c6deb4aab71808dcb408fe3d1e7690c43d37b4ce6cc0166fa7daa60808201526080815261089060a082610c96565b604051918291518091835e8101838152039060025afa15610b11575f519760205f604051828101907f03b03cd5effa95ac9bee94f1f5ef907157bda4812ccf0b4c91f42bb629f83a1c82527f1aa085ff28179a12d922dba0547057ccaae94b9d69cfaa4e60401fea7f3e033360408201527f110c10134f200b19f6490846d518c9aea868366efb7228ca5c91d2940d03076260608201527f1e60f31fcbf757e837e867178318832d0b2d74d59e2fea1c7142df187d3fc6d360808201526080815261095c60a082610c96565b604051918291518091835e8101838152039060025afa15610b11575f5160205f601d6040517f72697363305f67726f746831362e566572696679696e674b65792e4943000000815260025afa15610b11575f8051610140526101008190526060610120526020610160525b885180610100511015610b7a575f19810190808211610b66576101005190035f1901908111610b66578951811015610b5257610160519060051b8a0101519060405191610a176101205184610c96565b60028352610160516040903690850137610a3083610ccd565b52610a3a82610cda565b52604051610a4b6101605182610c96565b5f8152601f196101605101366101605183013781519061ffff8211610b3a5791604051928391610140516101605184015260408301815190916101605101905f905b808210610b1c575050509281610ad994600294935180926101605101825e019061ffff60f01b9061ff0060ff8260081c169160081b161760f01b16815203601d19810184520182610c96565b5f60405191805180916101605101845e820191818352806101605193039060025afa15610b11575f51610100805160010190526109c7565b6040513d5f823e3d90fd5b82518452610160518896509384019390920191600190910190610a8d565b506306dfcc6560e41b5f52601060045260245260445ffd5b634e487b7160e01b5f52603260045260245ffd5b634e487b7160e01b5f52601160045260245ffd5b505f92918b8b6040519661016051880195865260408801526060870152608086015260a085015260c0840152600560f81b8784015260c28352610bbe60e284610c96565b60405192518091845e820191818352806101605193039060025afa15610b11575f9182519060405194610160518601938452604086015260608501526080840152600360f81b60a084015260828352610c1860a284610c96565b60405192518091845e820191818352806101605193039060025afa15610b11575f516001600160e01b03191681526040516112129182610ceb83396080518281816105b90152610dc1015260a0518281816105740152610de7015260c0518281816101670152610e1f01525181818160ae0152610d2d0152f35b5f80fd5b601f909101601f19168101906001600160401b03821190821017610cb957604052565b634e487b7160e01b5f52604160045260245ffd5b805115610b525760200190565b805160011015610b52576040019056fe60806040526004361015610011575f80fd5b5f3560e01c8063053c238d146100945780631599ead51461008f578063258038e21461008a57806334baeab9146100855780638989fa2e146100805780639181e4b11461007b578063ab750e75146100765763ffa1ad7414610071575f80fd5b610703565b6105e9565b6105a4565b61055f565b6101a5565b610150565b6100db565b346100d7575f3660031901126100d75763ffffffff60e01b7f00000000000000000000000000000000000000000000000000000000000000001660805260206080f35b5f80fd5b346100d75760203660031901126100d7576004356001600160401b0381116100d75780360360406003198201126100d757600482013590602219018112156100d75781016004810135906001600160401b0382116100d7576024019080360382136100d757602461014e93013591610d29565b005b346100d7575f3660031901126100d75760206040517f00000000000000000000000000000000000000000000000000000000000000008152f35b906004916044116100d757565b9060c491610104116100d757565b346100d7576101a03660031901126100d7576101c03661018a565b3660c4116100d7576101d136610197565b366101a4116100d757604051906103808201604052610104356101f381610760565b610124359361020185610760565b6101443561020e81610760565b6101643561021b81610760565b610184359161022983610760565b60808701977f12ac9a25dcd5e1a832a9061a082c15dd1d61aa9c4d553505739d0f5d65dc3be4885260208801957f025aa744581ebe7ad91731911c898569106ff5a2d30f3eee2b23c60ee980acd487526102839089610791565b61028d908861081d565b61029790876108a9565b6102a19086610935565b6102ab90856109c1565b803585527f30644e72e131a029b85045b68181585d97816a916871ca8d3c208c16d87cfd4760209182013581030660a085015260443560c085015260643560e085015260843561010085015260a4356101208501527f2d4d9aa7e302d9df41749d5507949d05dbea33fbb16c643b22f599a2be6df2e26101408501527f14bedd503c37ceb061d8ec60209fe345ce89830a19230301f076caff004d19266101608501527f0967032fcbf776d1afc985f88877f182d38480a653f2decaa9794cbc3bf3060c6101808501527f0e187847ad4c798374d0d6732bf501847dd68bc0e071241e0213bc7fc13db7ab6101a08501527f304cfbd1e08a704a99f5e847d93f8c3caafddec46b7a0d379da69a4d112346a76101c08501527f1739c1b1a457a8c7313123d24d2f9192f896b7c63eea05a9d57f06547ad0cec86101e0850152835161020085015290516102208401527f198e9393920d483a7260bfb731fb5d25f1aa493335a9e71297e485b7aef312c26102408401527f1800deef121f1e76426a00665e5c4479674322d4f75edadd46debd5cd992f6ed6102608401527f090689d0585ff075ec9e99ad690c3395bc4b313370b38ef355acdadcd122975b6102808401527f12c85ea5db8c6deb4aab71808dcb408fe3d1e7690c43d37b4ce6cc0166fa7daa6102a084015281356102c084015201356102e08201527f03b03cd5effa95ac9bee94f1f5ef907157bda4812ccf0b4c91f42bb629f83a1c6103008201527f1aa085ff28179a12d922dba0547057ccaae94b9d69cfaa4e60401fea7f3e03336103208201527f110c10134f200b19f6490846d518c9aea868366efb7228ca5c91d2940d0307626103408201527f1e60f31fcbf757e837e867178318832d0b2d74d59e2fea1c7142df187d3fc6d36103609091015280806107cf195a01602092600861030092fa9051165f5260205ff35b346100d7575f3660031901126100d7576040517f00000000000000000000000000000000000000000000000000000000000000006001600160801b0319168152602090f35b346100d7575f3660031901126100d7576040517f00000000000000000000000000000000000000000000000000000000000000006001600160801b0319168152602090f35b346100d75760603660031901126100d7576004356001600160401b0381116100d757366023820112156100d7578060040135906001600160401b0382116100d75736602483830101116100d75761014e916024359060246044359301610a4d565b634e487b7160e01b5f52604160045260245ffd5b604081019081106001600160401b0382111761067957604052565b61064a565b60a081019081106001600160401b0382111761067957604052565b606081019081106001600160401b0382111761067957604052565b90601f801991011681019081106001600160401b0382111761067957604052565b604051906106e46040836106b4565b565b604051906106e460a0836106b4565b906106e460405192836106b4565b346100d7575f3660031901126100d75760405161071f8161065e565b6005815260406020820191640312e322e360dc1b83528151928391602083525180918160208501528484015e5f828201840152601f01601f19168101030190f35b7f30644e72e131a029b85045b68181585d2833e84879b9709143e1f593f0000001111561078957565b5f805260205ff35b604051917f0707b920bc978c02f292fae2036e057be54294114ccc3c8769d883f688a1423f83527f2e32a094b7589554f7bc357bf63481acd2d55555c203383782a4650787ff664260208401526040830190815260408360608160076107cf195a01fa1561078957815190526020810151606083015260409160809060066107cf195a01fa1561078957565b604051917f0bca36e2cbe6394b3e249751853f961511011c7148e336f4fd974644850fc34783527f2ede7c9acf48cf3a3729fa3d68714e2a8435d4fa6db8f7f409c153b1fcdf9b8b60208401526040830190815260408360608160076107cf195a01fa1561078957815190526020810151606083015260409160809060066107cf195a01fa1561078957565b604051917f1b8af999dbfbb3927c091cc2aaf201e488cbacc3e2c6b6fb5a25f9112e04f2a783527f2b91a26aa92e1b6f5722949f192a81c850d586d81a60157f3e9cf04f679cccd660208401526040830190815260408360608160076107cf195a01fa1561078957815190526020810151606083015260409160809060066107cf195a01fa1561078957565b604051917f2b5f494ed674235b8ac1750bdfd5a7615f002d4a1dcefeddd06eda5a076ccd0d83527f2fe520ad2020aab9cbba817fcbb9a863b8a76ff88f14f912c5e71665b2ad5e8260208401526040830190815260408360608160076107cf195a01fa1561078957815190526020810151606083015260409160809060066107cf195a01fa1561078957565b604051917f0f1c3c0d5d9da0fa03666843cde4e82e869ba5252fce3c25d5940320b1c4d49383527f214bfcff74f425f6fe8c0d07b307482d8bc8bb2f3608f68287aa01bd0b69e80960208401526040830190815260408360608160076107cf195a01fa1561078957815190526020810151606083015260409160809060066107cf195a01fa1561078957565b91610b02906106e4945f6080604051610a658161067e565b828152826020820152604051610a7a8161065e565b83815283602082015260408201528260608201520152610abb610a9b6106d5565b915f83525f6020840152610aad6106d5565b9081525f60208201526111a4565b90610ac46106e6565b9283527fa3acc27117418996340b84e5a90f3ef4c49d22c79e44aad822ec9c313e1eb8e2602084015260408301525f60608301526080820152610f5d565b91610d29565b906004116100d75790600490565b90929192836004116100d75783116100d757600401916003190190565b356001600160e01b0319811692919060048210610b4e575050565b6001600160e01b031960049290920360031b82901b16169150565b9080601f830112156100d75760405191610b846040846106b4565b8290604081019283116100d757905b828210610ba05750505090565b8135815260209182019101610b93565b610100818303126100d75760405191610bc883610699565b610bd28183610b69565b835280605f830112156100d7576040918251610bee84826106b4565b8060c08301928484116100d75785809101915b848310610c21575050506020850152610c1a9190610b69565b9082015290565b602090610c2e8785610b69565b8152019101908590610c01565b908160209103126100d7575180151581036100d75790565b905f905b60028210610c6457505050565b6020806001928551815201930191019091610c57565b905f905b60058210610c8b57505050565b6020806001928551815201930191019091610c7e565b919493929094610cb6836101a0810197610c53565b5f604084015b60028210610ce45750505081610cdd6101009260c06106e496950190610c53565b0190610c7a565b82515f90825b60028310610d08575050506020604060019201930191019091610cbc565b6020806001928451815201920192019190610cea565b6040513d5f823e3d90fd5b90917f0000000000000000000000000000000000000000000000000000000000000000610d6f610d62610d5c8686610b08565b90610b33565b6001600160e01b03191690565b6001600160e01b0319821603610ebc575090610da3610d9b84610d93602095611048565b969094610b16565b810190610bb0565b90610e5e82519160408585015194015195610dbe60a06106f5565b917f000000000000000000000000000000000000000000000000000000000000000060801c83527f000000000000000000000000000000000000000000000000000000000000000060801c8784015260801c604083015260801c60608201527f0000000000000000000000000000000000000000000000000000000000000000608082015260405195869485946334baeab960e01b865260048601610ca1565b0381305afa908115610eb7575f91610e88575b5015610e7957565b63439cc0cd60e01b5f5260045ffd5b610eaa915060203d602011610eb0575b610ea281836106b4565b810190610c3b565b5f610e71565b503d610e98565b610d1e565b610eef90610ecd610d5c8686610b08565b632e2ce35360e21b5f526001600160e01b031990811660045216602452604490565b5ffd5b60031115610efc57565b634e487b7160e01b5f52602160045260245ffd5b60205f60126040517172697363302e52656365697074436c61696d60701b815260025afa15610eb7575f5190565b516003811015610efc5790565b805191908290602001825e015f815290565b5f61103860209261102c610f6f610f10565b61101e606084015193805190888101519060406080820151910190610fc6610faa610fc08d610fb6610fa18751610f3e565b610faa81610ef2565b60181b63ff0000001690565b9551015160ff1690565b60ff1690565b604080518d8101988952602089019a909a52870194909452606086019290925260808501919091526001600160e01b031960e091821b811660a086015291901b1660a4830152600160fa1b60a8830152839160aa0190565b03601f1981018352826106b4565b60405191828092610f4b565b039060025afa15610eb7575f5190565b8060081c9060081b907cff000000ff000000ff000000ff000000ff000000ff000000ff000000ff7dff000000ff000000ff000000ff000000ff000000ff000000ff000000ff007fff000000ff000000ff000000ff000000ff000000ff000000ff000000ff00000084167eff000000ff000000ff000000ff000000ff000000ff000000ff000000ff000084161760101c931691161760101b1761110f7bffffffff00000000ffffffff00000000ffffffff00000000ffffffff8019831660201c921660201b90565b17604081811c77ffffffffffffffff0000000000000000ffffffffffffffff169177ffffffffffffffff0000000000000000ffffffffffffffff19911b161761116261115b8260801c90565b9160801b90565b17906111906111806111748460801c90565b6001600160801b031690565b60801b6001600160801b03191690565b60809290921b6001600160801b0319169190565b60205f600c6040516b1c9a5cd8cc0b93dd5d1c1d5d60a21b815260025afa15610eb7575f8051825160209384015160408051808701949094528301919091526060820152600160f91b6080820152606281526110389061102c6082826106b456fea164736f6c634300081a000a")]
+    contract RiscZeroGroth16Verifier {
+        constructor(bytes32 control_root, bytes32 bn254_control_id) {}
+    }
+}
