+++ crates/boundless-market/src/contracts/mod.rs
@@ -793,228 +793,7 @@ pub(crate) mod bytecode;
 #[cfg(feature = "test-utils")]
 #[allow(missing_docs)]
 /// Module for testing utilities.
-pub mod test_utils {
-    use crate::contracts::{
-        boundless_market::BoundlessMarketService,
-        bytecode::*,
-        hit_points::{default_allowance, HitPointsService},
-    };
-    use alloy::{
-        network::EthereumWallet,
-        node_bindings::AnvilInstance,
-        primitives::{Address, FixedBytes},
-        providers::{ext::AnvilApi, Provider, ProviderBuilder, WalletProvider},
-        signers::local::PrivateKeySigner,
-        sol_types::SolCall,
-    };
-    use anyhow::{Context, Result};
-    use risc0_ethereum_contracts::set_verifier::SetVerifierService;
-    use risc0_zkvm::sha::Digest;
-
-    pub struct TestCtx<P> {
-        pub verifier_addr: Address,
-        pub set_verifier_addr: Address,
-        pub hit_points_addr: Address,
-        pub boundless_market_addr: Address,
-        pub prover_signer: PrivateKeySigner,
-        pub customer_signer: PrivateKeySigner,
-        pub prover_provider: P,
-        pub prover_market: BoundlessMarketService<P>,
-        pub customer_provider: P,
-        pub customer_market: BoundlessMarketService<P>,
-        pub set_verifier: SetVerifierService<P>,
-        pub hit_points_service: HitPointsService<P>,
-    }
-
-    pub async fn deploy_mock_verifier<P: Provider>(deployer_provider: P) -> Result<Address> {
-        let instance = RiscZeroMockVerifier::deploy(deployer_provider, FixedBytes::ZERO)
-            .await
-            .context("failed to deploy RiscZeroMockVerifier")?;
-        Ok(*instance.address())
-    }
-
-    pub async fn deploy_set_verifier<P: Provider>(
-        deployer_provider: P,
-        verifier_address: Address,
-        set_builder_id: Digest,
-    ) -> Result<Address> {
-        let instance = RiscZeroSetVerifier::deploy(
-            deployer_provider,
-            verifier_address,
-            <[u8; 32]>::from(set_builder_id).into(),
-            String::default(),
-        )
-        .await
-        .context("failed to deploy RiscZeroSetVerifier")?;
-        Ok(*instance.address())
-    }
-
-    pub async fn deploy_hit_points<P: Provider>(
-        deployer_signer: &PrivateKeySigner,
-        deployer_provider: P,
-    ) -> Result<Address> {
-        let deployer_address = deployer_signer.address();
-        let instance = HitPoints::deploy(deployer_provider, deployer_address)
-            .await
-            .context("failed to deploy HitPoints contract")?;
-        Ok(*instance.address())
-    }
-
-    pub async fn deploy_boundless_market<P: Provider>(
-        deployer_signer: &PrivateKeySigner,
-        deployer_provider: P,
-        set_verifier: Address,
-        hit_points: Address,
-        assessor_guest_id: Digest,
-        allowed_prover: Option<Address>,
-    ) -> Result<Address> {
-        let deployer_address = deployer_signer.address();
-        let market_instance = BoundlessMarket::deploy(
-            &deployer_provider,
-            set_verifier,
-            <[u8; 32]>::from(assessor_guest_id).into(),
-            hit_points,
-        )
-        .await
-        .context("failed to deploy BoundlessMarket implementation")?;
-
-        let proxy_instance = ERC1967Proxy::deploy(
-            &deployer_provider,
-            *market_instance.address(),
-            BoundlessMarket::initializeCall {
-                initialOwner: deployer_address,
-                imageUrl: "".to_string(),
-            }
-            .abi_encode()
-            .into(),
-        )
-        .await
-        .context("failed to deploy BoundlessMarket proxy")?;
-        let proxy = *proxy_instance.address();
-
-        if hit_points != Address::ZERO {
-            let hit_points_service =
-                HitPointsService::new(hit_points, &deployer_provider, deployer_signer.address());
-            hit_points_service.grant_minter_role(hit_points_service.caller()).await?;
-            hit_points_service.grant_authorized_transfer_role(proxy).await?;
-            if let Some(prover) = allowed_prover {
-                hit_points_service.mint(prover, default_allowance()).await?;
-            }
-        }
-
-        Ok(proxy)
-    }
-
-    async fn deploy_contracts(
-        anvil: &AnvilInstance,
-        set_builder_id: Digest,
-        assessor_guest_id: Digest,
-    ) -> Result<(Address, Address, Address, Address)> {
-        let deployer_signer: PrivateKeySigner = anvil.keys()[0].clone().into();
-        let deployer_provider = ProviderBuilder::new()
-            .wallet(EthereumWallet::from(deployer_signer.clone()))
-            .on_builtin(&anvil.endpoint())
-            .await?;
-
-        // Deploy contracts
-        let verifier = deploy_mock_verifier(&deployer_provider).await?;
-        let set_verifier =
-            deploy_set_verifier(&deployer_provider, verifier, set_builder_id).await?;
-        let hit_points = deploy_hit_points(&deployer_signer, &deployer_provider).await?;
-        let boundless_market = deploy_boundless_market(
-            &deployer_signer,
-            &deployer_provider,
-            set_verifier,
-            hit_points,
-            assessor_guest_id,
-            None,
-        )
-        .await?;
-
-        // Mine forward some blocks using the provider
-        deployer_provider.anvil_mine(Some(10), Some(2)).await.unwrap();
-        deployer_provider.anvil_set_interval_mining(2).await.unwrap();
-
-        Ok((verifier, set_verifier, hit_points, boundless_market))
-    }
-
-    pub async fn create_test_ctx(
-        anvil: &AnvilInstance,
-        set_builder_id: impl Into<Digest>,
-        assessor_guest_id: impl Into<Digest>,
-    ) -> Result<TestCtx<impl Provider + WalletProvider + Clone + 'static>> {
-        create_test_ctx_with_rpc_url(anvil, &anvil.endpoint(), set_builder_id, assessor_guest_id)
-            .await
-    }
-
-    pub async fn create_test_ctx_with_rpc_url(
-        anvil: &AnvilInstance,
-        rpc_url: &str,
-        set_builder_id: impl Into<Digest>,
-        assessor_guest_id: impl Into<Digest>,
-    ) -> Result<TestCtx<impl Provider + WalletProvider + Clone + 'static>> {
-        let (verifier_addr, set_verifier_addr, hit_points_addr, boundless_market_addr) =
-            deploy_contracts(anvil, set_builder_id.into(), assessor_guest_id.into()).await.unwrap();
-
-        let prover_signer: PrivateKeySigner = anvil.keys()[1].clone().into();
-        let customer_signer: PrivateKeySigner = anvil.keys()[2].clone().into();
-        let verifier_signer: PrivateKeySigner = anvil.keys()[0].clone().into();
-
-        let prover_provider = ProviderBuilder::new()
-            .wallet(EthereumWallet::from(prover_signer.clone()))
-            .on_builtin(rpc_url)
-            .await?;
-        let customer_provider = ProviderBuilder::new()
-            .wallet(EthereumWallet::from(customer_signer.clone()))
-            .on_builtin(rpc_url)
-            .await?;
-        let verifier_provider = ProviderBuilder::new()
-            .wallet(EthereumWallet::from(verifier_signer.clone()))
-            .on_builtin(rpc_url)
-            .await?;
-
-        let prover_market = BoundlessMarketService::new(
-            boundless_market_addr,
-            prover_provider.clone(),
-            prover_signer.address(),
-        );
-
-        let customer_market = BoundlessMarketService::new(
-            boundless_market_addr,
-            customer_provider.clone(),
-            customer_signer.address(),
-        );
-
-        let set_verifier = SetVerifierService::new(
-            set_verifier_addr,
-            verifier_provider.clone(),
-            verifier_signer.address(),
-        );
-
-        let hit_points_service = HitPointsService::new(
-            hit_points_addr,
-            verifier_provider.clone(),
-            verifier_signer.address(),
-        );
-
-        hit_points_service.mint(prover_signer.address(), default_allowance()).await?;
-
-        Ok(TestCtx {
-            verifier_addr,
-            set_verifier_addr,
-            hit_points_addr,
-            boundless_market_addr,
-            prover_signer,
-            customer_signer,
-            prover_provider,
-            prover_market,
-            customer_provider,
-            customer_market,
-            set_verifier,
-            hit_points_service,
-        })
-    }
-}
+pub mod test_utils;
 
 #[cfg(test)]
 mod tests {
