+++ crates/order-generator/src/main.rs
@@ -250,7 +250,10 @@ mod tests {
     use alloy::{
         node_bindings::Anvil, providers::Provider, rpc::types::Filter, sol_types::SolEvent,
     };
-    use boundless_market::contracts::{test_utils::create_test_ctx, IBoundlessMarket};
+    use boundless_market::contracts::{
+        test_utils::{create_test_ctx, DEV_MODE_TRUE},
+        IBoundlessMarket,
+    };
     use guest_assessor::ASSESSOR_GUEST_ID;
     use guest_set_builder::SET_BUILDER_ID;
     use tracing_test::traced_test;
@@ -261,7 +264,9 @@ mod tests {
     #[traced_test]
     async fn test_main() {
         let anvil = Anvil::new().spawn();
-        let ctx = create_test_ctx(&anvil, SET_BUILDER_ID, ASSESSOR_GUEST_ID).await.unwrap();
+        let ctx = create_test_ctx(&anvil, SET_BUILDER_ID, ASSESSOR_GUEST_ID, DEV_MODE_TRUE)
+            .await
+            .unwrap();
 
         let args = MainArgs {
             rpc_url: anvil.endpoint_url(),
