+++ crates/slasher/tests/basic.rs
@@ -12,7 +12,8 @@ use alloy::{
     signers::Signer,
 };
 use boundless_market::contracts::{
-    test_utils::create_test_ctx, Input, Offer, Predicate, PredicateType, ProofRequest, Requirements,
+    test_utils::{create_test_ctx, DEV_MODE_TRUE},
+    Input, Offer, Predicate, PredicateType, ProofRequest, Requirements,
 };
 use futures_util::StreamExt;
 use guest_assessor::ASSESSOR_GUEST_ID;
@@ -56,7 +57,8 @@ async fn create_order(
 async fn test_basic_usage() {
     let anvil = Anvil::new().spawn();
     let rpc_url = anvil.endpoint_url();
-    let ctx = create_test_ctx(&anvil, SET_BUILDER_ID, ASSESSOR_GUEST_ID).await.unwrap();
+    let ctx =
+        create_test_ctx(&anvil, SET_BUILDER_ID, ASSESSOR_GUEST_ID, DEV_MODE_TRUE).await.unwrap();
 
     let exe_path = env!("CARGO_BIN_EXE_boundless-slasher");
     let args = [
